ojyamaに関することを新しいファイルに移した｡ ゲームオーバーになる条件を別の関数として新しいファイルに移した
